requirements.txt
torch>=2.0
torch-geometric
networkx
nbformat
nbconvert
biopython
requests
beautifulsoup4
pronto
neo4j
spacy
vllm
langchain
langgraph
pydantic
